feat(soc): the Orin SM's sourced FP16 tensor rate (#269 5.3) (#318)
nvidia_ampere_sm_orin gains fp16: 2048 dense ops per SM per clock, half
the INT8 rate, from two NVIDIA sources that agree:
- Jetson AGX Orin Series Modules data sheet DS-10662-001 v1.8 (Sept 2025),
  JAO 64GB: 'Up to 170 INT8 Sparse TOPS or 85 Sparse FP16 TFLOPS (Tensor
  Cores)' -> 42.5 dense / (16 SM x 1.3 GHz) = 2043.
- NVIDIA Technical Blog 'Jetson Orin Nano Developer Kit Gets a Super Boost',
  Table 1: 8 SM, 33 TOPS (Dense) INT8 and 17 FP16 TFLOPs at 1,020 MHz ->
  2083; 20 dense and 10 at 635 MHz -> 1969. The unlabelled Nano FP16
  figure is dense: it is half the dense INT8 figure at the same clock.

Consequences:
- orin_nano_measured_v1 re-ingested: the GPU FP16 measurements become four
  CALIBRATED entries (attention 0.57, conv/GEMM 0.48, layer norm, GEMV) --
  27 entries in all.
- Class B on the Orin GPU runs in FP16 rather than falling back to FP32.
  In air superiority det is no longer over and the GPU bottleneck drops
  from 1.53 to 1.09.
- No node states FP16 energy per op (BF16 is a different format), so
  those Class B ops are a dynamic-power gap on real data and TOPS/W is
  withheld; tests that exercise the priced path use a node copy with
  test-only FP16 energy, and a test pins the real gap.
- show_soc prints the FP16 peak; docs that said the SM had no FP16 rate
  are updated.

+++ tests/estimation/test_soc_schedule.py
@@ -172,10 +172,10 @@ def test_service_time_is_ops_over_peak_times_efficiency_per_class(orin):
     svc = engine_service(demand, KernelClass.DENSE_CONV_GEMM, gpu, _everything_known(0.5), 0.0)
     a, b, _ = demand.stage.class_split
     ops = demand.stage.ops_per_call
-    # Class A in INT8; Class B in FP32 because the SM template has no FP16.
-    assert svc.formats == {"A": "int8", "B": "fp32"}
+    # Class A in INT8; Class B in FP16, the SM's sourced tensor rate.
+    assert svc.formats == {"A": "int8", "B": "fp16"}
     expected = ops * a / (gpu.server_peak_ops_per_s("int8") * 0.5) + ops * b / (
-        gpu.server_peak_ops_per_s("fp32") * 0.5)
+        gpu.server_peak_ops_per_s("fp16") * 0.5)
     assert svc.t_compute_s == pytest.approx(expected)
     assert svc.t_memory_s is None and svc.bound == "compute"
